Delete README.md
Large libraries resorted to automating the checkout process as it’s impossible to keep track of all their customers, and what they’ve borrowed. Our project aims to implement a miniature model of how a typical library checkout system which is completely digitized works using Arduino Uno and RFID including features such as students borrowing and returning books, and viewing student and book details, all with little to no human intervention. Our project uses RFID for student authentication and also for unique student and book identification. From this project, we have not only learned about the applications of Arduino with RFID but also how it can be used with Python. We use Python with SQLite for our database. Combining the beautiful features of both Python and Arduino, we make a seamless interface for students, making a student’s library checkout process fast and efficient.